reorganiza classes cliente
