+++ src/main/java/br/com/joaolira/desafiojava1/cliente/application/api/ClienteAPI.java
@@ -1,5 +1,10 @@
 package br.com.joaolira.desafiojava1.cliente.application.api;
 
+import br.com.joaolira.desafiojava1.cliente.application.api.request.ClienteAlteracaoRequest;
+import br.com.joaolira.desafiojava1.cliente.application.api.request.ClienteRequest;
+import br.com.joaolira.desafiojava1.cliente.application.api.response.ClienteDetalhadoResponse;
+import br.com.joaolira.desafiojava1.cliente.application.api.response.ClienteListResponse;
+import br.com.joaolira.desafiojava1.cliente.application.api.response.ClienteResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
