+++ src/main/java/br/com/joaolira/desafiojava1/endereco/application/api/EnderecoAPI.java
@@ -1,6 +1,5 @@
 package br.com.joaolira.desafiojava1.endereco.application.api;
 
-import br.com.joaolira.desafiojava1.cliente.application.api.ClienteAlteracaoRequest;
 import br.com.joaolira.desafiojava1.endereco.application.api.request.EnderecoAlteracaoRequest;
 import br.com.joaolira.desafiojava1.endereco.application.api.request.EnderecoRequest;
 import br.com.joaolira.desafiojava1.endereco.application.api.response.DetalhaEnderecoResponse;
